requirements.txt
pandas
numpy
scikit-learn
xgboost
pyyaml
mlflow
fastapi
uvicorn
pydantic
joblib
pytest
httpx
python-multipart
python-dotenv